+++ OpenGLStudy/Engine/Graphics/EnvironmentCaptureManager.h
@@ -16,10 +16,10 @@ namespace Graphics
 			cSphere BV;								// The bounding volume of this probe, should not change after initialization; 
 			cSphere InnerBV;						// If a sphere is inside the inner sphere, this weight should be 1
 			float Influence;								// 1 at the inner boundary; 0 at the outer boundary; > 1 inside the inner boundary. < 0, outside the outer boundary
-			float Resolution;							// Resolution of Environment cubemap(each size)
+			GLuint Resolution;							// Resolution of Environment cubemap(each size)
 
 			sCaptureProbes() {}
-			sCaptureProbes(const cSphere& i_sphere, const cSphere& i_innerSphere, float i_influence, float i_resolution) : BV(i_sphere), InnerBV(i_innerSphere), Influence(i_influence), Resolution(i_resolution) { }
+			sCaptureProbes(const cSphere& i_sphere, const cSphere& i_innerSphere, float i_influence, GLuint i_resolution) : BV(i_sphere), InnerBV(i_innerSphere), Influence(i_influence), Resolution(i_resolution) { }
 			sCaptureProbes(const sCaptureProbes& i_other) :
 				EnvironmentProbe(i_other.EnvironmentProbe), IrradianceProbe(i_other.IrradianceProbe), PrefilterProbe(i_other.PrefilterProbe),
 				BV(i_other.BV), InnerBV(i_other.InnerBV), Influence(i_other.Influence), Resolution(i_other.Resolution) {}
